requirements.txt
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
iniconfig==2.0.0
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy==1.1.1
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pluggy==1.2.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytest==7.4.0
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
